towards a new capped impl; NamespaceDetails readability; cleaning

+++ db/extsort.cpp
@@ -218,14 +218,13 @@ namespace mongo {
     // -----------------------------------
     
     BSONObjExternalSorter::FileIterator::FileIterator( string file ){
-        long length;
+        unsigned long long length;
         _buf = (char*)_file.map( file.c_str() , length , MemoryMappedFile::SEQUENTIAL );
         massert( 10308 ,  "mmap failed" , _buf );
-        assert( (unsigned long long)length == (unsigned long long)file_size( file ) );
+        assert( length == file_size( file ) );
         _end = _buf + length;
     }
-    BSONObjExternalSorter::FileIterator::~FileIterator(){
-    }
+    BSONObjExternalSorter::FileIterator::~FileIterator() {}
     
     bool BSONObjExternalSorter::FileIterator::more(){
         return _buf < _end;
